Use proper int for main() function

+++ src/doors/ny2008/src/nyibbs.cpp
@@ -608,7 +608,7 @@ void            RefreshBest(void) {
 
 
 
-INT16             main(INT16 argc, char *argv[]) {
+int             main(int argc, char *argv[]) {
 	FILE           *justfile, *njustfile;
 	//char          key;
 	//WORD  uintval;
